Hide widgets if process isn't editable.

+++ src/lib/ARCI.svelte
@@ -1,6 +1,6 @@
 <!-- Represents ARCI for a processes's how to -->
 <script lang="ts">
-	import type { HowRow, ProcessRow } from '$database/OrganizationsDB';
+	import type { HowRow } from '$database/OrganizationsDB';
 	import Button, { Delete } from './Button.svelte';
 	import Level from './Level.svelte';
 	import RoleLink from './RoleLink.svelte';
@@ -9,6 +9,7 @@
 
 	export let how: HowRow;
 	export let verbose: boolean;
+	export let editable: boolean;
 
 	const org = getOrg();
 	const db = getDB();
@@ -42,7 +43,7 @@
 		{#if !verbose}
 			<Level level="responsible" />
 		{/if}
-		{#if options.length > 0}
+		{#if options.length > 0 && editable}
 			<Select
 				tip="Add a role to be responsible for completing this step."
 				{options}
@@ -58,10 +59,12 @@
 		{/if}
 		{#each how.responsible as responsible}
 			<RoleLink roleID={responsible} />
-			<Button
-				tip="Remove this role from the responsible list."
-				action={() => $db.removeHowRCI(how, responsible, 'responsible')}>{Delete}</Button
-			>
+			{#if editable}
+				<Button
+					tip="Remove this role from the responsible list."
+					action={() => $db.removeHowRCI(how, responsible, 'responsible')}>{Delete}</Button
+				>
+			{/if}
 		{/each}
 		{#if verbose}
 			is <Level level="responsible" verbose /> for completing this process.
@@ -71,7 +74,7 @@
 		{#if !verbose}
 			<Level level="consulted" />
 		{/if}
-		{#if options.length > 0}
+		{#if options.length > 0 && editable}
 			<Select
 				tip="Add a role to be consulted in this step."
 				{options}
@@ -87,10 +90,12 @@
 		{/if}
 		{#each how.consulted as consulted}
 			<RoleLink roleID={consulted} />
-			<Button
-				tip="Remove this role from the consulted list."
-				action={() => $db.removeHowRCI(how, consulted, 'consulted')}>{Delete}</Button
-			>
+			{#if editable}
+				<Button
+					tip="Remove this role from the consulted list."
+					action={() => $db.removeHowRCI(how, consulted, 'consulted')}>{Delete}</Button
+				>
+			{/if}
 		{/each}
 		{#if verbose}
 			is <Level level="consulted" verbose /> to complete this process.
@@ -100,7 +105,7 @@
 		{#if !verbose}
 			<Level level="informed" />
 		{/if}
-		{#if options.length > 0}
+		{#if options.length > 0 && editable}
 			<Select
 				tip="Add a role to be informed in this step."
 				{options}
@@ -116,10 +121,12 @@
 		{/if}
 		{#each how.informed as informed}
 			<RoleLink roleID={informed} />
-			<Button
-				tip="Remove this role from the informed list"
-				action={() => $db.removeHowRCI(how, informed, 'informed')}>{Delete}</Button
-			>
+			{#if editable}
+				<Button
+					tip="Remove this role from the informed list"
+					action={() => $db.removeHowRCI(how, informed, 'informed')}>{Delete}</Button
+				>
+			{/if}
 		{/each}
 		{#if verbose}
 			is <Level level="informed" verbose /> about this process and its outcomes.
